[System] Use Supercronic for cron
With using Fly.io and containers, it doesn't like using cron, so supercronic takes care of tasks, does mean we lose the niceness of whenever, but does mean we can at least run the tasks as expected.

Manually having to set up a cronfile for now, potentially will need to find some ways to use whenever later on, for now this works well.